refactor: improve CLI structure and enhance pre-commit hook functionality
- Cleaned up whitespace in `src/cyt/cli/app.py` and `src/cyt/proxy/cli.py` for better readability.
- Updated `DEV.md` to clarify the separation of `cyt-client` and `cyt-mcp` scripts.
- Enhanced `pyproject.toml` and `typos.toml` to include additional ignored words for better typo management.
- Improved `prek-loop.sh` and `prek-hook-groups.yaml` to support parallel execution of pre-commit hooks, enhancing development efficiency.
- Refactored various functions in `src/cyt/hook/http_server.py` and `src/cyt/tiers` modules for better clarity and maintainability.
- This update aims to streamline the development workflow and improve the overall structure of the CLI and pre-commit hooks.

+++ src/cyt/tiers/scores.py
@@ -24,7 +24,7 @@ def wilson_lower_bound(successes: float, trials: float, *, z: float = 1.96) -> f
 def decay_factor(*, requests_since: int, half_life: float) -> float:
     if requests_since <= 0 or half_life <= 0:
         return 1.0
-    return 2.0 ** (-requests_since / half_life)
+    return float(2.0 ** (-requests_since / half_life))
 
 
 def demand_score(stats: object) -> float:
